tests: たまにunitsテストが落ちるバグ修正
