+++ src/features/editor/extensions/nodes/message/tests/parser.spec.ts
@@ -1,0 +1,37 @@
+import Document from "@tiptap/extension-document";
+import Paragraph from "@tiptap/extension-paragraph";
+import Text from "@tiptap/extension-text";
+import { describe, expect, it } from "vitest";
+import { createEditorInstance } from "@/tests/editor-instance";
+import { Message } from "../message";
+import { MessageContent } from "../message-content";
+
+describe("HTMLのパース", () => {
+  it("aside.msgをメッセージノードとしてパースできる", () => {
+    const editor = createEditorInstance({
+      extensions: [Document, Paragraph, Text, Message, MessageContent],
+      content:
+        '<aside class="msg"><div class="msg-content"><p>メッセージ</p></div></aside>',
+    });
+
+    const docString = editor.state.doc.toString();
+
+    expect(docString).toBe(
+      'doc(message(messageContent(paragraph("メッセージ"))))',
+    );
+  });
+
+  it("aside.msg.alertをアラートタイプとしてパースできる", () => {
+    const editor = createEditorInstance({
+      extensions: [Document, Paragraph, Text, Message, MessageContent],
+      content:
+        '<aside class="msg alert"><div class="msg-content"><p>メッセージ</p></div></aside>',
+    });
+
+    const docString = editor.state.doc.toString();
+
+    expect(docString).toBe(
+      'doc(message(messageContent(paragraph("メッセージ"))))',
+    );
+  });
+});
